implementing date format, adding currency setting 18/2/2025

+++ app/Models/Purchase.php
@@ -40,7 +40,7 @@ class Purchase extends Model
     }
 
     protected $casts = [
-        'total' => 'double',
+        'total' => 'float',
         'order_date' => 'datetime',
         'due_date' => 'datetime',
         'updated_at' => 'datetime'
